renamed reservation_manager files to hotel_booker, adjusted spec_helper
