+++ lib/hotel_booker.rb
