+++ spec/hotel_booker_spec.rb
