+++ spec/spec_helper.rb
@@ -17,4 +17,4 @@ Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 require_relative '../lib/reservation'
 require_relative '../lib/date_range'
-require_relative '../lib/reservation_manager'
+require_relative '../lib/hotel_booker'
